pyproject.toml
[project]
name = "XSerialOne"
version = "0.1.0"
description = "Xbox controller interface via custom serial XInputOne device"
authors = [{"name" = "Ellie Vallard", "email" = ""}]
license = {"text" = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pyserial",
    "pygame",
    "numpy",
]

[project.urls]
Homepage = "https://github.com/Elisiah/XSerialOne"
Repository = "https://github.com/Elisiah/XSerialOne.git"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E501",   # line too long: some generated/data lines exceed 120; enforce manually
    "E701",   # multiple statements on one line: existing style in several modules
    "E741",   # ambiguous variable name: common in math/input code
    "F601",   # repeated dict key: exists in frame constant mappings
]